pyproject.toml
[project]
name = "fast-agent-mcp"
version = "0.1.12"
description = "Define, Prompt and Test MCP enabled Agents and Workflows"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Shaun Smith", email = "fastagent@llmindset.co.uk"},
    { name = "Sarmad Qadri", email = "sarmad@lastmileai.dev" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent"
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.6",
    "mcp==1.6.0",
    "opentelemetry-distro>=0.50b0",
    "opentelemetry-exporter-otlp-proto-http>=1.29.0",
    "pydantic-settings>=2.7.0",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "typer>=0.15.1",
    "numpy>=2.2.1",
    "scikit-learn>=1.6.0",
    "anthropic>=0.49.0",
    "openai>=1.63.2",
    "prompt-toolkit>=3.0.50",
    "aiohttp>=3.11.13",
]

[project.optional-dependencies]
temporal = [
    "temporalio>=1.8.0",
]
openai = [
    "openai>=1.58.1",
]
dev = [
    "anthropic>=0.42.0",
    "pre-commit>=4.0.1",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "ruff>=0.8.4",
    "tomli>=2.2.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_agent"]

[tool.hatch.build]
include = [
    "src/mcp_agent/**/*.py",
    "src/mcp_agent/**/*.yaml",
    "src/mcp_agent/**/*.yml",
    "src/mcp_agent/**/*.csv",
    "src/mcp_agent/resources/**/*.py",
    "src/mcp_agent/resources/**/*.yaml",
    "src/mcp_agent/resources/**/*.yml",
    "src/mcp_agent/resources/**/*.csv",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "e2e: tests that connect to external resources (llms)",
    "integration: marks tests as integration tests",
    "simulated_endpoints: marks tests that use simulated external endpoints"
]
# Other pytest options can go here too
testpaths = ["test", "integration_tests"]

[dependency-groups]
dev = [
    "anthropic>=0.49.0",
    "pre-commit>=4.0.1",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "ruff>=0.8.4",
    "tomli>=2.2.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
]

[project.scripts]
fast-agent = "mcp_agent.cli.__main__:app"
fast_agent = "mcp_agent.cli.__main__:app"
fastagent = "mcp_agent.cli.__main__:app"
silsila = "mcp_agent.cli.__main__:app"
prompt-server = "mcp_agent.mcp.prompts.__main__:main"

[tool.setuptools.package-data]
mcp_agent = [
    "data/*.json",
    "resources/examples/**/*.py",
    "resources/examples/**/*.yaml",
    "resources/examples/**/*.yml",
    "resources/examples/**/*.csv",
    "resources/examples/**/mount-point/*.csv"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "ANN", # type annotations
    "TCH"  # type checking
]

# More lenient settings for existing code
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]  # Don't require type annotations in tests
